refactor jsp
